1.4.2.102 source update from zip

+++ Projectiles/Melee/StarCrystal.cs
@@ -35,7 +35,7 @@ namespace CalamityMod.Projectiles.Melee
 	            {
 	            	Dust.NewDust(projectile.position + projectile.velocity, projectile.width, projectile.height, 227, projectile.oldVelocity.X * 0.5f, projectile.oldVelocity.Y * 0.5f);
 	            	Projectile.NewProjectile(projectile.position.X, projectile.position.Y, (float)Main.rand.Next(-35, 36) * 0.2f, (float)Main.rand.Next(-35, 36) * 0.2f, mod.ProjectileType("TinyCrystal"), 
-	            	(int)((double)projectile.damage * 0.5), (float)((int)((double)projectile.knockBack * 0.15)), Main.myPlayer, 0f, 0f);
+	            	(int)((double)projectile.damage * 0.5), projectile.knockBack * 0.15f, Main.myPlayer, 0f, 0f);
 	            }
         	}
         }
